test: Merged SVG tests into test_generate.py..
